sage-free/requirements.txt
# Core ML & Inference
vllm==0.6.0
torch==2.4.0
transformers==4.44.0
accelerate==0.33.0
sentence-transformers==3.0.1
faiss-cpu==1.8.0
gudhi==3.10.1

# Orchestration & Logic
langgraph==0.2.14
langchain==0.2.14
langchain-openai==0.1.22
pydantic==2.8.2
numpy==1.26.4
scipy==1.13.1

# API & UI
fastapi==0.112.0
uvicorn==0.30.6
gradio==4.40.0
python-multipart==0.0.9

# Utilities
python-dotenv==1.0.1
httpx==0.27.0
pyyaml==6.0.1
ruff==0.6.1
pytest==8.3.2
pytest-asyncio==0.23.8
loguru==0.7.2
